+++ src/fetch.ts
@@ -1,3 +1,4 @@
+import TypeValidator from './type.js';
 import { Static, TSchema, TUnknown } from "@sinclair/typebox";
 import { TypeCompiler } from "@sinclair/typebox/compiler";
 import { fetch, Response } from 'undici';
@@ -16,16 +17,7 @@ export class TypedResponse extends Response {
 
     async typed<T extends TSchema = TUnknown>(type: T): Promise<Static<T>> {
         const body = await this.json();
-
-        const typeChecker = TypeCompiler.Compile(type)
-        const result = typeChecker.Check(body);
-
-        if (result) return body;
-
-        const errors = typeChecker.Errors(body);
-        const firstError = errors.First();
-
-        throw new Error(`Internal Validation Error: ${JSON.stringify(firstError)} -- Body: ${JSON.stringify(body)}`);
+        return TypeValidator.type(type, body);
     }
 }